Create single_threading.py
This is an example of a single-threaded program. It has a task called job, which sleeps for three seconds as well as outputs text. We note the time, then run that job function twice, then note the time again, ready to output how long everything took in total. As each job run takes three seconds, when we run the job twice, it will report it took a total of six seconds