Create multi_threading.py
So what did we change here? Well, first we're importing threading from the Thread library to help give us multithreading capabilities. Then we start two threads using the Thread class, and in initializing that class, provide the target function and the args arguments (which are a tuple, hence the brackets and commas).

We then start those threads by using the start() method, and we also decide to use join() to join the new threads to the main thread; so, it has to wait for those threads to complete in the main thread